+++ components/filemanager/impl/src/main/java/com/flipperdevices/filemanager/impl/composable/ComposableFileManagerScreen.kt
@@ -16,8 +16,10 @@ import com.flipperdevices.deeplink.api.DeepLinkParser
 import com.flipperdevices.deeplink.model.DeeplinkContent
 import com.flipperdevices.filemanager.impl.R
 import com.flipperdevices.filemanager.impl.composable.bar.ComposableFileManagerTopBar
+import com.flipperdevices.filemanager.impl.composable.dialog.ComposableInputDialog
 import com.flipperdevices.filemanager.impl.composable.dialog.ComposableSelectDialog
 import com.flipperdevices.filemanager.impl.composable.list.ComposableFileManagerContent
+import com.flipperdevices.filemanager.impl.model.CreateFileManagerAction
 import com.flipperdevices.filemanager.impl.model.FileItem
 import com.flipperdevices.filemanager.impl.model.FileManagerState
 import com.flipperdevices.filemanager.impl.viewmodels.FileManagerViewModel
@@ -57,6 +59,14 @@ fun ComposableFileManagerScreen(
         }
     }
 
+    var showAddDialog by remember { mutableStateOf(false) }
+    if (showAddDialog) {
+        ComposableCreateActionDialog(
+            onCreateAction = fileManagerViewModel::onCreateAction,
+            onDismiss = { showAddDialog = false }
+        )
+    }
+
     ComposableFileManagerScreenInternal(
         fileManagerState = fileManagerState,
         deepLinkParser = deepLinkParser,
@@ -67,8 +77,52 @@ fun ComposableFileManagerScreen(
         },
         onUploadFile = {
             onUploadFile(fileManagerState.currentPath, it)
+        },
+        onAddButton = {
+            showAddDialog = true
         }
     )
+}
+
+@Composable
+private fun ComposableCreateActionDialog(
+    onCreateAction: (CreateFileManagerAction, String) -> Unit,
+    onDismiss: () -> Unit
+) {
+    var createFileManagerAction by remember { mutableStateOf<CreateFileManagerAction?>(null) }
+    val localCreateFileManagerAction = createFileManagerAction
+
+    if (localCreateFileManagerAction != null) {
+        ComposableInputDialog(
+            when (localCreateFileManagerAction) {
+                CreateFileManagerAction.FILE -> R.string.add_dialog_title_file
+                CreateFileManagerAction.FOLDER -> R.string.add_dialog_title_folder
+            }
+        ) {
+            if (it != null) {
+                onCreateAction(localCreateFileManagerAction, it)
+            }
+            onDismiss()
+            createFileManagerAction = null
+        }
+    }
+
+    ComposableSelectDialog(
+        intArrayOf(
+            R.string.filemanager_add_dialog_file,
+            R.string.filemanager_add_dialog_folder
+        )
+    ) {
+        when (it) {
+            R.string.filemanager_add_dialog_file -> {
+                createFileManagerAction = CreateFileManagerAction.FILE
+            }
+            R.string.filemanager_add_dialog_folder -> {
+                createFileManagerAction = CreateFileManagerAction.FOLDER
+            }
+            else -> onDismiss()
+        }
+    }
 }
 
 @Composable
@@ -76,7 +130,8 @@ private fun ComposableFileManagerScreenInternal(
     fileManagerState: FileManagerState,
     onOpenFolder: (FileItem) -> Unit,
     deepLinkParser: DeepLinkParser,
-    onUploadFile: (DeeplinkContent) -> Unit
+    onUploadFile: (DeeplinkContent) -> Unit,
+    onAddButton: () -> Unit
 ) {
     val context = LocalContext.current
 
@@ -95,9 +150,13 @@ private fun ComposableFileManagerScreenInternal(
 
     Scaffold(
         topBar = {
-            ComposableFileManagerTopBar(fileManagerState.currentPath) {
-                pickFileLauncher.launch("*/*")
-            }
+            ComposableFileManagerTopBar(
+                fileManagerState.currentPath,
+                onClickUploadButton = {
+                    pickFileLauncher.launch("*/*")
+                },
+                onClickAddButton = onAddButton
+            )
         }
     ) { scaffoldPaddings ->
         ComposableFileManagerContent(