+++ components/filemanager/impl/src/main/java/com/flipperdevices/filemanager/impl/composable/bar/ComposableFileManagerTopBar.kt
@@ -1,5 +1,6 @@
 package com.flipperdevices.filemanager.impl.composable.bar
 
+import androidx.compose.foundation.layout.Row
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
 import androidx.compose.material.TopAppBar
@@ -10,7 +11,11 @@ import com.flipperdevices.core.ui.res.R as DesignSystem
 import com.flipperdevices.filemanager.impl.R
 
 @Composable
-fun ComposableFileManagerTopBar(path: String, onClickUploadButton: () -> Unit) {
+fun ComposableFileManagerTopBar(
+    path: String,
+    onClickUploadButton: () -> Unit,
+    onClickAddButton: () -> Unit
+) {
     TopAppBar(
         title = {
             ComposableEllipsizeStartText(
@@ -19,15 +24,27 @@ fun ComposableFileManagerTopBar(path: String, onClickUploadButton: () -> Unit) {
         },
         actions = {
             if (isAbleToSafe(path)) {
-                IconButton(onClick = onClickUploadButton) {
-                    Icon(
-                        painter = painterResource(
-                            DesignSystem.drawable.ic_upload
-                        ),
-                        contentDescription = stringResource(
-                            R.string.filemanager_upload_action
+                Row {
+                    IconButton(onClick = onClickAddButton) {
+                        Icon(
+                            painter = painterResource(
+                                DesignSystem.drawable.ic_plus
+                            ),
+                            contentDescription = stringResource(
+                                R.string.filemanager_create_action
+                            )
                         )
-                    )
+                    }
+                    IconButton(onClick = onClickUploadButton) {
+                        Icon(
+                            painter = painterResource(
+                                DesignSystem.drawable.ic_upload
+                            ),
+                            contentDescription = stringResource(
+                                R.string.filemanager_upload_action
+                            )
+                        )
+                    }
                 }
             }
         }